Change user_params being passed to #create
The user_params that were being defined within their own method of
the same name were using require and permit for username, email,
and password, but password was not getting passed along to the
creation of a new User therefore each new user entered doing
registration was invalid and not persisted in the database. To fix
this the individual params calls for username, email, and password
were passed directly to the User.create inside the user create
action of the users controller.