Remove some validations for user model
New user instances were not being created because the User model
called for the presence of their city, state, zip, and bio
information which is generally not passed through on initial
registration. These attributes will be added by the user after their
initial registration, therefore the presence_of validation for city,
state, zip, and bio were removed from the model. The username
attribute is also actually supposed to be the user's actual name and
since different people can have the same name, the uniqueness
validation for username was also removed.