Change some validations
There were too many validations going on for the user inside of its
model. We had previously decided that upon registration (user creation)
that the user would not be entering there city, state, zipcode, or bio
and that these attributes would be edited by the user at another time.
Thus, the validations for the presence of the attributes upon creation
were unnecessary and kept throwing errors for registration inside the
app since these attributes were not included in the registration form.