Minor syntax changes
Most of the "differences" shown between the two versions of the file
are small line changes and syntax changes such as changing root_url
to root_path inside the private method titled logged_in_user. Another
change was that the attribute password_confirmation was added to the
require/permit for params inside of the private user_params method to
be used in the create action. There was a problem with the create
action since the controller called for a redirect to an html page,
which is not handled/accepted by the React Native app which only
accepts JSON. Therefore the redirect_to @user was changed to
render json: @user

- #logged_in_user: root_url -> root_path
- #user_params: params.require("").permit("", :password_confirmation)
- #create: redirect_to @user -> render json: @user